Add initial version of stimulus setup script.
This is an initial version that is working but missing some features that we definitely want. Main ones relate to parametrisation of randomisation (which here is fixed: with replacement, no order effects controlled). This has not been tested on Windows so would be great if it could be tested (and please try to break the input sanitisation!).